separated dom methods
