+++ packages/dom/src/index.ts
@@ -1,3 +1,5 @@
+import { allInputsSelector } from './allInputsSelector';
+import { isNativeInput } from './isNativeInput';
 import { append } from './append';
 import { blockElements } from './blockElements';
 import { calculateBaseline } from './calculateBaseline';
@@ -7,7 +9,7 @@ import { fragmentToString } from './fragmentToString';
 import { getContentLength } from './getContentLength';
 import { getDeepestBlockElements } from './getDeepestBlockElements';
 import { getDeepestNode } from './getDeepestNode';
-import { allInputsSelector, isNativeInput, findAllInputs } from './inputs';
+import { findAllInputs } from './findAllInputs';
 import { isCollapsedWhitespaces } from './isCollapsedWhitespaces';
 import { isContentEditable } from './isContentEditable';
 import { isElement } from './isElement';
@@ -16,7 +18,8 @@ import { isFragment } from './isFragment';
 import { isHTMLString } from './isHtmlString';
 import { isLeaf } from './isLeaf';
 import { isNodeEmpty } from './isNodeEmpty';
-import { isSingleTag, isLineBreakTag } from './isTag';
+import { isLineBreakTag } from './isLineBreakTag';
+import { isSingleTag } from './isSingleTag';
 import { make } from './make';
 import { offset } from './offset';
 import { prepend } from './prepend';
