+++ packages/dom/src/isSingleTag.ts
@@ -23,15 +23,3 @@ export function isSingleTag(tag: HTMLElement): boolean {
     'WBR',
   ].includes(tag.tagName);
 }
-
-/**
- * Check if element is BR or WBR
- * @param element - element to check
- * @returns boolean that represents if element is a line break tag
- */
-export function isLineBreakTag(element: HTMLElement): element is HTMLBRElement {
-  return [
-    'BR',
-    'WBR',
-  ].includes(element.tagName);
-}
